offchain/prover/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hashcredit-prover"
version = "0.1.0"
description = "HashCredit Bitcoin SPV Proof Builder - generates Merkle proofs for payout verification"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "HashCredit Team"}
]
keywords = ["bitcoin", "spv", "merkle", "proof", "hashcredit"]

dependencies = [
    "web3>=6.0.0",
    "httpx>=0.25.0",
    "eth-account>=0.10.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "typer>=0.9.0",
    "structlog>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
hashcredit-prover = "hashcredit_prover.cli:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
